% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-test.R
\name{annotation_test}
\alias{annotation_test}
\title{Annotation: statistical test}
\usage{
annotation_test(
  mapping = NULL,
  data = NULL,
  method = NULL,
  formula = NULL,
  label_x,
  label_y,
  hjust = 0.5,
  vjust = 0.5,
  method_args = NULL,
  tidy_fn = NULL,
  ...,
  angle = 0,
  units = "npc",
  label_gp = NULL,
  na.rm = FALSE,
  orientation = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[=aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping. This is not the same with other \code{ggplot2} geoms, which is just for
statistical test but not for plot.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[=ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[=fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{method}{the method used to implement test.}

\item{formula}{passed to function specified in \strong{method}.}

\item{label_x, label_y}{the coordinates of the label, usually in "npc" units.}

\item{hjust, vjust}{The justification of the text relative to its (x, y)
location. Possible string values are: "left", "right", "centre", "center",
"bottom", and "top". For numeric values, 0 means left (bottom) alignment and
1 means right (top) alignment. \strong{hjust} for horizontal justification and
\strong{vjust} for vertical justification.}

\item{method_args}{other arguments passed to function specified in
\strong{method}.}

\item{tidy_fn}{A function or formula which accepts results returned by
function in \strong{method} and return a scalar character.

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function with up to
two arguments: \code{.x} (single argument) or \code{.x} and \code{.y} (two arguments). The
\code{.} placeholder can be used instead of \code{.x}.  This allows you to create
very compact anonymous functions (lambdas) with up to two inputs. Functions
created from formulas have a special class. Use \code{is_lambda()} to test for
it.

If a \strong{string}, the function is looked up in \code{globalenv()}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{ggplot2::layer()}}. These are often
aesthetics, used to set an aesthetic to a fixed value, like \code{colour = "red"} or \code{size = 3}. They may also be parameters to the paired geom/stat.}

\item{angle}{The angle to rotate the text.}

\item{units}{A string indicating the default units \code{labelx} and \code{label_y}
use.}

\item{label_gp}{other arguments to define the text attributes.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a
warning. If \code{TRUE}, missing values are silently removed.}

\item{orientation}{The orientation of the layer. The default (‘NA’)
automatically determines the orientation from the aesthetic mapping.
In the rare event that this fails it can be given explicitly by setting
'orientation' to either "x" or "y"}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
Most useful for adding statistical test results.
}
