% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-comparetest.R, R/stat-comparetest.R
\name{geom_comparetest}
\alias{geom_comparetest}
\alias{stat_comparetest}
\title{Statistical Compare test}
\usage{
geom_comparetest(
  mapping = NULL,
  data = NULL,
  stat = "comparetest",
  position = "identity",
  height = ggplot2::rel(0.05),
  step_increase = 1,
  tip_length = ggplot2::rel(0.01),
  ...,
  nudge_x = NULL,
  nudge_y = NULL,
  parse = FALSE,
  arrow = NULL,
  arrow_fill = NULL,
  lineend = "butt",
  linejoin = "round",
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)

stat_comparetest(
  mapping = NULL,
  data = NULL,
  method = NULL,
  compare_list = NULL,
  method_args = NULL,
  tidy_fn = NULL,
  geom = "comparetest",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{height}{numeric vector indicating the value where label start. use
\link[ggplot2:element]{rel} to signal values as the fraction of maximal height}

\item{step_increase}{numeric vector indicating the increase for every
additional comparison to minimize overlap, use \link[ggplot2:element]{rel} to signal
values as the fraction of maximal height .}

\item{tip_length}{numeric vector indicating the length of the beard, use
\link[ggplot2:element]{rel} to signal values as the fraction of the maximal height}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{nudge_x, nudge_y}{Horizontal and vertical adjustment to nudge labels by.}

\item{parse}{If \code{TRUE}, the labels will be parsed into expressions and
displayed as described in \code{?plotmath}.}

\item{arrow}{specification for arrow heads, as created by \code{\link[grid:arrow]{grid::arrow()}}.}

\item{arrow_fill}{fill colour to use for the arrow head (if closed). \code{NULL}
means use \code{colour} aesthetic.}

\item{lineend}{Line end style (round, butt, square).}

\item{linejoin}{Line join style (round, mitre, bevel).}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with a warning.
If \code{TRUE} silently removes missing values.}

\item{orientation}{The orientation of the layer. The default (‘NA’)
automatically determines the orientation from the aesthetic mapping.
In the rare event that this fails it can be given explicitly by setting
'orientation' to either "x" or "y"}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{method}{the method used to implement test. Defaults to "nonparametric",
which will use "wilcox.test" to compare two groups and use "kruskal.test" to
compare three or more groups. Support lamda created from a formula. This can
also be "none" indicates no test will be implemented.}

\item{compare_list}{A list of atomic vectors with at least 2 length. The
entries in the vector are the values on the x-axis indicating the comparison
among what groups.}

\item{method_args}{arguments passed to function specified in \strong{method}.}

\item{tidy_fn}{A function or formula which accepts results returned by
function in \strong{method} and return a scalar character.

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function with up to
two arguments: \code{.x} (single argument) or \code{.x} and \code{.y} (two arguments). The
\code{.} placeholder can be used instead of \code{.x}.  This allows you to create
very compact anonymous functions (lambdas) with up to two inputs. Functions
created from formulas have a special class. Use \code{is_lambda()} to test for
it.

If a \strong{string}, the function is looked up in \code{globalenv()}.
Notes: if \strong{method} is "none", this can be a list (whose length should equal
to the number of \strong{PANEL}) of labels corresponding to each comparisons in
compare_list, this can be matched by position or names if both \strong{tidy_fn}
and \strong{compare_list} have names.}

\item{geom}{The geometric object to use to display the data, either as a
\code{ggproto} \code{Geom} subclass or as a string naming the geom stripped of the
\code{geom_} prefix (e.g. \code{"point"} rather than \code{"geom_point"})}
}
\description{
Statistical Compare test
}
\section{Computed variables}{

\code{stat_comparetest()} provides the following variables, some of which depend on the orientation:
\describe{
\item{xmin \emph{or} ymin}{the left (or lower) side of horizontal (or vertical)
segments underneath label}
\item{xmax \emph{or} ymax}{the right (or upper) side of horizontal (or vertical)
segments underneath label}
\item{x \emph{or} y}{the x (or y) coordinates for labels, usually equal to
(xmin + xmax) / 2 or (ymin + ymax) / 2}
\item{y \emph{or} x}{the y (or x) coordinates for labels, usually equal to
(ymin + ymax) / 2 or (xmin + xmax) / 2}
\item{label}{the statistical test results}
\item{tip}{a list of data.frame gives the coordinates of tip where x
corresponds to the x axis of current comparison group, and y corresponds to
the maximal values of current comparison group the tip length is reverse to
the y value}
}
}

