% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{ggforest}
\alias{ggforest}
\title{Forest plot in ggplot2}
\usage{
ggforest(
  data,
  left_table = NULL,
  right_table = NULL,
  ylabels = NULL,
  nudge_y = 0.5,
  y_labels_nudge = 0.5,
  y_labels_position = "left",
  point_shape = 16L,
  point_size = 4L,
  point_color = "darkred",
  null_line_at = 0L,
  null_linetype = "dashed",
  null_line_params = list(),
  errorbar_width = 0.15,
  errorbar_params = list(),
  xlab = NULL,
  xlim = NULL,
  xbreaks = waiver(),
  xlabels = scales::number_format(accuracy = 0.1),
  x_scale_trans = "log10",
  x_scale_expand = c(0, 0),
  y_scale_expand = c(0, 0),
  left_table_params = list(),
  right_table_params = list(),
  add_arrow = TRUE,
  arrow_labels = c("Lower", "Higher"),
  arrow_weights = 35L,
  grid_arrow = grid::arrow(angle = 15, type = "closed", length = grid::unit(0.1, "in")),
  add_band = TRUE,
  band_col = c("white", "#eff3f2"),
  forest_width = 5L,
  widths = NULL,
  heights = c(30L, 1L)
)
}
\arguments{
\item{data}{A matrix or data.frame-like data. Excluding the \code{ylabels} column,
the first three column will be regarded as the forest point estimate and
confidence interval (lower and higher).}

\item{left_table, right_table}{Add a table in the left or right, details see
\link{ggtable}.}

\item{ylabels}{A string, in \code{colnames(data)}, specifies the column used as
the y-axis text.}

\item{nudge_y}{Vertical adjustment to nudge all errorbar or table text by.}

\item{y_labels_nudge}{Vertical adjustment in [0, 1] of y-axis labels.}

\item{y_labels_position}{The position of the y-axis labels, left or right for
\code{y_labels_position}.}

\item{point_shape, point_size, point_color}{The shape, size, color of the
point estimate.}

\item{null_line_at, null_linetype}{Add a vertical line in the
forest plot to indiate null hypothesis.}

\item{null_line_params}{Other argument passed to
\link[ggplot2:geom_abline]{geom_vline}.}

\item{errorbar_width}{Width of errorbar. Passed to
\link[ggplot2:geom_linerange]{geom_errorbar}.}

\item{errorbar_params}{Other arguments passed to
\link[ggplot2:geom_linerange]{geom_errorbar}.}

\item{xlab, xlim, xbreaks, xlabels, x_scale_trans, x_scale_expand}{Arguments
passed as \code{name}, \code{limits}, \code{breaks}, \code{labels}, \code{trans}, \code{expand} in
\link[ggplot2:scale_continuous]{scale_x_continuous}. Just a note: use xlab =
\code{expression(Protective \%<->\% Hazardous)} to add directed arrow in x labs. But
in this function, we just assembly an arrow vertically.}

\item{y_scale_expand}{A vector of range expansion
constants used to add some padding around the data to ensure that they are
placed some distance away from the axes. Will also control the
\code{y_scale_expand} of \link{ggtable}.}

\item{left_table_params, right_table_params}{Other arguments passed to
\link{ggtable}.}

\item{add_arrow}{A bool, if \code{TRUE}, will add arrow below x-axis.}

\item{arrow_labels}{String Vector, length 2. Labels for the arrows. Set
arrows to TRUE or this will have no effect.}

\item{arrow_weights}{A numeric of length 2, reduce this if you want to
shorten the arrow, and increase this if you want to lengthen the arrow.}

\item{grid_arrow}{An \link[grid:arrow]{arrow} object created.}

\item{add_band}{A bool, if \code{TRUE}, will add band interval in the background.}

\item{band_col}{Color of background band interval.}

\item{forest_width}{The relative width of the forest plot, only used when
\code{widths} is \code{NULL}. The relative width of talbe will be calculated as the
column number.}

\item{widths, heights}{The relative widths and heights of each column and row
in the grid. See \link[patchwork:wrap_plots]{wrap_plots}.}
}
\value{
A \link[ggplot2:ggplot]{ggplot} object or
\link[patchwork:wrap_plots]{patchwork}.
}
\description{
Forest plot in ggplot2
}
