% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggstattest-package.R
\docType{package}
\name{ggstattest-package}
\alias{ggstattest}
\alias{ggstattest-package}
\title{ggstattest: GGplot2 extension for statistical test}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Yun Peng \email{yunyunpp96@outlook.com}

}
\keyword{internal}
