% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{ggtable}
\alias{ggtable}
\title{Graphical display of a textual table}
\usage{
ggtable(
  data,
  cols = NULL,
  ylabels = NULL,
  mapping = aes(),
  nudge_x = 0.5,
  nudge_y = 0.5,
  hjust = 0.5,
  vjust = 0.5,
  x_scale_expand = c(0L, 0L),
  y_scale_expand = c(0L, 0L),
  x_labels_nudge = 0.5,
  x_labels_position = NULL,
  x_labels_element = NULL,
  y_labels_nudge = 0.5,
  y_labels_position = NULL,
  y_labels_element = NULL,
  clip = "off",
  add_band = TRUE,
  band_col = c("white", "#eff3f2")
)
}
\arguments{
\item{data}{A matrix or data.frame-like data.}

\item{cols}{Columns used to create table. If \code{NULL}, will select all columns
except column specified by \code{ylabels}.}

\item{ylabels}{Extract a single column used as the y-axis text, passed to
\code{var} argument in \link[dplyr:pull]{pull}.}

\item{mapping}{List of aesthetic mappings to use for plot, created by \link{aes}.
This will be added into the default \link[ggplot2:ggplot]{ggplot} object to
change the default aesthetic mappings. The column selected in \code{cols} will be
collected in \code{.__value__.} with column names in \code{.__name__.}. This will be
helpful if you want to adjust aesthetic by specific column or row.}

\item{nudge_x, nudge_y}{Horizontal and vertical adjustment [0, 1] to nudge
text by in a table cell.}

\item{hjust, vjust}{Horizontal and vertical justification [0, 1] of text in
each cell of table.}

\item{x_scale_expand, y_scale_expand}{A vector of range expansion
constants used to add some padding around the data to ensure that they are
placed some distance away from the axes. See
\link[ggplot2:scale_continuous]{scale_continuous}.}

\item{x_labels_nudge, y_labels_nudge}{Horizontal and vertical adjustment in
[0, 1] of table header.}

\item{x_labels_position, y_labels_position}{The position of the header,
left or right for \code{y_labels_position}, top or bottom for
\code{x_labels_position}.}

\item{x_labels_element, y_labels_element}{A
\link[ggplot2:element]{element_text} object specifies the header attributes.}

\item{clip}{Should drawing be clipped to the extent of the plot panel? A
setting of "on" means yes, and a setting of "off" (the default) means no.}

\item{add_band}{A bool, if \code{TRUE}, will add band interval in the background.}

\item{band_col}{Color of background band interval.}
}
\value{
A \link[ggplot2:ggplot]{ggplot} object.
}
\description{
Create a table in ggplot2 style
}
