% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{ggtable}
\alias{ggtable}
\title{Graphical display of a textual table}
\usage{
ggtable(
  data,
  x = 0.5,
  y = 0.5,
  hjust = 0.5,
  vjust = 0.5,
  row_header = 0.5,
  row_header_text = NULL,
  col_header = 0.5,
  column_header_text = NULL,
  add_band = TRUE,
  band_col = c("white", "#eff3f2")
)
}
\arguments{
\item{data}{A matrix or data.frame-like data.}

\item{x, y}{Axis value in [0, 1] of text in each cell of table.}

\item{hjust, vjust}{Horizontal and vertical justification [0, 1] of text in
each cell of table.}

\item{row_header, col_header}{Axis value in [0, 1] of table header.}

\item{row_header_text, column_header_text}{A
\link[ggplot2:element]{element_text} object specifies the header attributes.}

\item{add_band}{A bool, if \code{TRUE}, will add band interval in the background.}

\item{band_col}{Color of background band interval.}
}
\value{
A \link[ggplot2:ggplot]{ggplot} object.
}
\description{
Create a table in ggplot2 style
}
